Try to use project AVA if available
Closes https://github.com/avajs/atom-ava/issues/19
Previously the package would use the globally installed AVA. According
to a comment on that same issue, global AVA takes care of finding the
closest local AVA installation. But if you don't have any global
packages, then that would never work and would fail.

With this commit, when you're on a given file in the Atom editor, it
will use Atom's API to figure out the project path of that
file/folder, and from there it will check if that project has the AVA
binary under `project-path/node_modules/.bin/ava`. If so, it will use
that. If any of those checks fail, it will default back to the global
AVA.