Add tests for using the project AVA behaviour
These tests cover the new behaviour of atom-ava using the project AVA
when available. There's a bit more info on a comment for the new
tests, but for these we need to mock a call made to the Atom API, and
the file check made with fs so we trick the test environment into
thinking we're running inside an Atom project, and that the project
AVA binary exists so it will generate the project AVA path instead of
using the global one.